Fix issue with call log last modified not being updated.
In the past a security update was done which replaced usage of the
DbModifierWithNotification class with SQLiteQueryBuilders.
This solved the security issue but caused all the code in the
DbModifierWithNotification update and delete methods to no longer run.
This includes notifying of new written voicemails, updating the last
modified date for rows, and notifying content observers.
We cleaned up the content observer issue in QPR, so it was time to swing
back and clean the rest up.

In the end it just makes more sense to use SQLiteQueryBuilder instances
in DbModifierWithNotification for the user-provided queries; this lets
the rest of the update code which resided in that class still operate
in the same manner as it used to.

Test: Run unit tests
Test: Run CTS tests
Test: Add new CTS test to verify call log last modified gets updated.
Fixes: 177421643
Change-Id: Idc73df91cb1a0f4e77a4c5d2a288ab6ff2f6fc77
Merged-In: I3cddf913dec54f40358de095659673d4623518db
(cherry picked from commit 7f70b88e9759b2cdbe3c18a8321df569bbb8d509)
(cherry picked from commit 0ac4a3c0f45e31469898a22a283cc2dc1b44ee6d)
(cherry picked from commit b9ac3ae341b833211577990805806a95199893ef)